+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ### Changed
 - `esbuild` plugin: use `jsr:@deno/loader` official package instead of `@luca/esbuild-deno-loader`.
 - Middlewares configured in the `server.middlewares` option are registered before the dev middlewares.
+- `--watch` and `--serve` modes don't exit if there's an error on build.
 
 ### Fixed
 - Optimized file server by removing a `Deno.stat` call.